Update Gateway_Ethernet.ino
Added MQTT subscription, parsing, outgoing transmit to nodes and parsing on the node side. This sketch currently subscribes to 4 topics, 1111, 2222, 3333 and 4444. A MQTT message sent to those topics will be picked up by this sketch. If the message is formatted node#,device#,value# essentially 3 numbers separated by a comma, then the sketch will transmit the entire payload to the node matching the first number. Nodes configured with a different node number and not in promiscuous mode will not receive the message. Once received by the node, the payload is converted into three integers, "node", "device" and "value", from which you should be able to do whatever you want.